requirements.txt

Django==3.1.6
psycopg2==2.8.6
django-db-logger==0.1.10   
django-tracking2==0.5.0    
djangorestframework==3.12.2
Pillow==8.1.0
xhtml2pdf
boto3
pytz
PyPDF2
